fix(chat): rebase in-ride history integration onto refreshed feature/463
Squashes the previous two #474 commits (e7c96d28, 5509d741) and rebases
them onto the post-merge feature/463 so the diff against main only carries
the in-ride persistence + history rehydration work:

- TripChatProcessor::processInRide forwards GPS position and POI payload
  to persistChatTurn so the JSONB column and geo_* columns are filled.
- TripChatHistoryProvider rehydrates POIs into typed PoiSuggestionDto
  instances; repository exposes findHistory with cursor pagination and
  DEFAULT_LIMIT/MAX_LIMIT constants.
- Entity exposes the precise array shape for pois so PHPStan can prove
  the PoiSuggestionDto mapping.
- distance_m/detour_m use int across the wire (rounded meters), matching
  PoiSuggestion::toArray().
- Zod safeUrlSchema validates deeplink as http(s) before render.

+++ api/src/ApiResource/Model/PoiSuggestionDto.php
@@ -29,10 +29,10 @@ final readonly class PoiSuggestionDto
         public float $lat,
         #[ApiProperty(description: 'POI longitude (WGS84).', required: true)]
         public float $lon,
-        #[ApiProperty(description: 'Straight-line distance from the rider to the POI, in meters.', required: true)]
-        public float $distance_m,
-        #[ApiProperty(description: 'Estimated additional meters if the rider detours to the POI (0 when no remaining route is known).', required: true)]
-        public float $detour_m,
+        #[ApiProperty(description: 'Straight-line distance from the rider to the POI, in meters (rounded).', required: true)]
+        public int $distance_m,
+        #[ApiProperty(description: 'Estimated additional meters if the rider detours to the POI (0 when no remaining route is known, rounded).', required: true)]
+        public int $detour_m,
         #[ApiProperty(description: 'Raw OSM `opening_hours` tag for the current day, when available.')]
         public ?string $opening_hours_today,
         #[ApiProperty(description: 'RFC 3339 closing time of the currently-open interval, or null when the POI never closes / is closed.')]
@@ -52,8 +52,8 @@ final readonly class PoiSuggestionDto
      *     category: string,
      *     lat: float,
      *     lon: float,
-     *     distance_m: float,
-     *     detour_m: float,
+     *     distance_m: int,
+     *     detour_m: int,
      *     opening_hours_today: ?string,
      *     closes_at: ?string,
      *     phone: ?string,
